Delete clean.go
It did not make sense to have a separate file called clean.go since the method is of Cache struct so put back the clean() method to cache.go
