+++ cache.go
@@ -51,3 +51,19 @@ func (c *Cache) Find(key string) (interface{}, bool) {
 	go c.clean()
 	return r.value, true
 }
+
+// deletes expired records
+func (c *Cache) clean() {
+  foundExpired := false
+	for k, v := range c.recordMap {
+		if v.hasExpired() {
+      if !foundExpired{
+        // Lock only if an entry for expiration has been found
+        foundExpired = true
+        cMutex.Lock()
+        defer cMutex.Unlock()
+      }
+			delete(c.recordMap, k)
+		}
+	}
+}
